added setcohesitystoragedomain

+++ src/Cohesity.Powershell.Models/DatastoreInfo.cs
@@ -36,7 +36,7 @@ namespace Cohesity.Models
         /// </summary>
         /// <param name="capacity">Specifies the capacity of the datastore in bytes..</param>
         /// <param name="freeSpace">Specifies the available space on the datastore in bytes..</param>
-        public DatastoreInfo(int? capacity = default(int?), int? freeSpace = default(int?))
+        public DatastoreInfo(long? capacity = default(long?), long? freeSpace = default(long?))
         {
             this.Capacity = capacity;
             this.FreeSpace = freeSpace;
@@ -47,14 +47,14 @@ namespace Cohesity.Models
         /// </summary>
         /// <value>Specifies the capacity of the datastore in bytes.</value>
         [DataMember(Name="capacity", EmitDefaultValue=false)]
-        public int? Capacity { get; set; }
+        public long? Capacity { get; set; }
 
         /// <summary>
         /// Specifies the available space on the datastore in bytes.
         /// </summary>
         /// <value>Specifies the available space on the datastore in bytes.</value>
         [DataMember(Name="freeSpace", EmitDefaultValue=false)]
-        public int? FreeSpace { get; set; }
+        public long? FreeSpace { get; set; }
 
         /// <summary>
         /// Returns the string presentation of the object
